Separated syntatic sugar from ADTManagar
