+++ Source/ProofKitLib/adtmanager.swift
@@ -21,8 +21,6 @@ public struct ADTManager{
     self["set"] = Set()
     self["sequence"] = Sequence()
     self["int"] = Integer()
-    // self["petrinet"] = Petrinet()
-    // self["marking"] = Marking()
   }
 
   public subscript ( i : String) -> ADT{
@@ -142,74 +140,4 @@ public struct ADTManager{
 
     return "?"
   }
-}
-precedencegroup SuchThatPrecendence {
-  lowerThan: LogicalConjunctionPrecedence
-}
-
-precedencegroup GEvalPrecendece{
-  higherThan: SuchThatPrecendence, LogicalConjunctionPrecedence
-  lowerThan: AdditionPrecedence
-}
-
-
-// infix operator ∧: LogicalConjunctionPrecedence
-
-// infix operator ∨: LogicalConjunctionPrecedence
-infix operator <->: GEvalPrecendece
-infix operator ∈: AdditionPrecedence
-infix operator =>: SuchThatPrecendence
-
-public func =>(left: @escaping Goal, right: @escaping Goal)-> Goal{
-  return left && right
-}
-
-public func <->(left: Term, right: Term)-> Goal  {
-  return ADTm.geval(operation: left,result: right)
-}
-
-public func ∈(left: Term, right: ADT.Type)->Goal{
-  return right.belong(left)
-}
-
-public func <(left: Term, right: ADT.Type)->Goal{
-  return right.belong(left)
-}
-
-public func +(left: Term, right: Term)-> Term{
-  return Operator.n("+", left, right)
-}
-
-public func -(left: Term, right: Term)-> Term{
-  return Operator.n("+", left, right)
-}
-
-public func *(left: Term, right: Term)-> Term{
-  return Operator.n("*", left, right)
-}
-
-public func /(left: Term, right: Term)-> Term{
-  return Operator.n("/", left, right)
-}
-
-
-public func >(left: Term, right: Term)-> Term{
-  return Operator.n(">", left, right)
-}
-
-public func <(left: Term, right: Term)-> Term{
-  return Operator.n("<", left, right)
-}
-
-public func ==(left: Term, right: Term)-> Term{
-  return Operator.n("==", left, right)
-}
-
-public func ||(left: Term, right: Term)-> Term{
-  return Operator.n("or", left, right)
-}
-
-
-public func &&(left: Term, right: Term)-> Term{
-  return Operator.n("and", left, right)
 }
